--- lang-common/src/generator.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod test_generator {}
