+++ lang-common/src/type_defines.rs
@@ -1,0 +1,2 @@
+pub mod generators;
+pub mod type_define;
