+++ lang-common/src/types/froms.rs
@@ -1,0 +1,1 @@
+pub mod from_json;
